\name{partition_data_partitionfinder}
\alias{partition_data_partitionfinder}
\title{
partition data from a PartitionFinder output file
}
\description{
Partition the data from a partition finder output file
}
\usage{
partition_data_partitionfinder(data_file_fasta, partition_finder_output)
}
\arguments{
  \item{data_file_fasta}{
Name of the file of concatenated data in fasta format
}
}

\value{
This function does not return anything on the prompt. It saves the output files from a partitionfinder output file into the workind directory.
}

\author{
Sebastian Duchene
}

\seealso{
#Pending.
}
\examples{

\dontrun{
Pending.
}
}
\keyword{ sBSDmin }
